Readable timer values added.

+++ app/src/main/java/com/oolong/countdown_timer/utils/StartCountdownService.kt
@@ -7,7 +7,7 @@ import androidx.annotation.RequiresApi
 import com.oolong.countdown_timer.services.CountdownService
 
 @RequiresApi(Build.VERSION_CODES.O)
-fun Context.startCountdownService(action: String, duration: Int) {
+fun Context.startCountdownService(action: String, duration: Long) {
     Intent(this, CountdownService::class.java).also {
         it.action = action
         it.putExtra("duration", duration)
